refactor: split useFeedbackData into own module, remove unused item prop
Address architecture review feedback:
- Move useFeedbackData hook from feedback-utils.ts into use-feedback-data.ts
  so the utils module stays pure (no React Query coupling)
- Remove unused `item` prop from FeedbackActions component
